style: move hooks
